Task 1 almost done
